+++ R/indicators.R
@@ -18,7 +18,9 @@
 #' @seealso \code{\link{area_types}} for area type  and their parent mappings,
 #'   \code{\link{indicator_metadata}} for indicator metadata and
 #'   \code{\link{profiles}} for profile lookups and
-#'   \code{\link{deprivation_decile}} for deprivation decile lookups
+#'   \code{\link{deprivation_decile}} for deprivation decile lookups and
+#'   \code{\link{category_types}} for category lookups and
+#'   \code{\link{indicator_areatypes}} for indicators by area types lookups
 #' @export
 
 indicators <- function(ProfileID = NULL,